+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - New validator for html lang tag (shortcut: lang)
 - New validator for viewport tag (shortcut: vp)
+- New validator for descriptive link texts (shortcut: atxt)
 
 ## [0.2]